포트폴리오 & 프로덕션 querydsl 수정 커밋

+++ src/main/java/com/tsp/new_tsp_project/api/admin/model/controller/AdminModelJpaApi.java
@@ -67,7 +67,7 @@ public class AdminModelJpaApi {
 		ConcurrentHashMap modelMap = searchCommon.searchCommon(page, paramMap);
 		modelMap.put("categoryCd", categoryCd);
 
-		Integer modelListCnt = this.adminModelJpaService.findModelsCount(modelMap);
+		Long modelListCnt = this.adminModelJpaService.findModelsCount(modelMap);
 
 		List<AdminModelDTO> modelList = new ArrayList<>();
